lab + inventory

+++ app/Http/Controllers/ClientController.php
@@ -9,14 +9,21 @@ use Illuminate\Http\Request;
 class ClientController extends Controller
 {
     public function destroy(Request $request , Client $client){
+        $user =  auth()->user();
+        if ($user->can('حذف عميل')) {
         return ['status'=>$client->delete()];
+        }else{
+            return response(['status' => false,'message'=>'صلاحيه حذف العميل غير مفعله'],400);
+
+        }
     }
     public  function index(){
         return Client::all();
     }
     public function create(ClientFormRequest $request){
 //        $request->validate($request->all());
-
+        $user =  auth()->user();
+        if ($user->can('اضافه عميل')) {
         $data = $request->all();
 //        return $data;
        $result = Client::create([
@@ -26,7 +33,11 @@ class ClientController extends Controller
             'email'=>$data['email'],
 
         ]);
-       return ['status'=> $result->id];
+       return ['status'=> true, 'data' => $result];
+        }else{
+            return response(['status' => false,'message'=>'صلاحيه اضافه العميل غير مفعله'],400);
+
+        }
     }
     public function edit(Request $request , Client $client){
         $data = $request->all();
